move OffHeapVectorValues to its own class
