+++ lucene/core/src/java/org/apache/lucene/codecs/lucene91/Lucene91HnswVectorsReader.java
@@ -20,7 +20,6 @@ package org.apache.lucene.codecs.lucene91;
 import static org.apache.lucene.search.DocIdSetIterator.NO_MORE_DOCS;
 
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
@@ -32,7 +31,6 @@ import org.apache.lucene.index.FieldInfo;
 import org.apache.lucene.index.FieldInfos;
 import org.apache.lucene.index.IndexFileNames;
 import org.apache.lucene.index.RandomAccessVectorValues;
-import org.apache.lucene.index.RandomAccessVectorValuesProducer;
 import org.apache.lucene.index.SegmentReadState;
 import org.apache.lucene.index.VectorSimilarityFunction;
 import org.apache.lucene.index.VectorValues;
@@ -281,7 +279,7 @@ public final class Lucene91HnswVectorsReader extends KnnVectorsReader {
     IOUtils.close(vectorData, vectorIndex);
   }
 
-  private static class FieldEntry {
+  static class FieldEntry {
 
     final VectorSimilarityFunction similarityFunction;
     final long vectorDataOffset;
@@ -291,7 +289,7 @@ public final class Lucene91HnswVectorsReader extends KnnVectorsReader {
     final int maxConn;
     final int numLevels;
     final int dimension;
-    private final int size;
+    final int size;
     final int[][] nodesByLevel;
     // for each level the start offsets in vectorIndex file from where to read neighbours
     final long[] graphOffsetsByLevel;
@@ -583,80 +581,6 @@ public final class Lucene91HnswVectorsReader extends KnnVectorsReader {
     Bits getAcceptOrds(Bits acceptDocs) {
       return null;
     }
-  }
-
-  /** Read the vector values from the index input. This supports both iterated and random access. */
-  abstract static class OffHeapVectorValues extends VectorValues
-      implements RandomAccessVectorValues, RandomAccessVectorValuesProducer {
-
-    protected final int dimension;
-    protected final int size;
-    protected final IndexInput slice;
-    protected final BytesRef binaryValue;
-    protected final ByteBuffer byteBuffer;
-    protected final int byteSize;
-    protected final float[] value;
-
-    OffHeapVectorValues(int dimension, int size, IndexInput slice) {
-      this.dimension = dimension;
-      this.size = size;
-      this.slice = slice;
-      byteSize = Float.BYTES * dimension;
-      byteBuffer = ByteBuffer.allocate(byteSize);
-      value = new float[dimension];
-      binaryValue = new BytesRef(byteBuffer.array(), byteBuffer.arrayOffset(), byteSize);
-    }
-
-    @Override
-    public int dimension() {
-      return dimension;
-    }
-
-    @Override
-    public int size() {
-      return size;
-    }
-
-    @Override
-    public long cost() {
-      return size;
-    }
-
-    @Override
-    public float[] vectorValue(int targetOrd) throws IOException {
-      slice.seek((long) targetOrd * byteSize);
-      slice.readFloats(value, 0, value.length);
-      return value;
-    }
-
-    @Override
-    public BytesRef binaryValue(int targetOrd) throws IOException {
-      readValue(targetOrd);
-      return binaryValue;
-    }
-
-    private void readValue(int targetOrd) throws IOException {
-      slice.seek((long) targetOrd * byteSize);
-      slice.readBytes(byteBuffer.array(), byteBuffer.arrayOffset(), byteSize);
-    }
-
-    public abstract int ordToDoc(int ord);
-
-    private static OffHeapVectorValues load(FieldEntry fieldEntry, IndexInput vectorData)
-        throws IOException {
-      if (fieldEntry.docsWithFieldOffset == -2) {
-        return new EmptyOffHeapVectorValues(fieldEntry.dimension);
-      }
-      IndexInput bytesSlice =
-          vectorData.slice("vector-data", fieldEntry.vectorDataOffset, fieldEntry.vectorDataLength);
-      if (fieldEntry.docsWithFieldOffset == -1) {
-        return new DenseOffHeapVectorValues(fieldEntry.dimension, fieldEntry.size, bytesSlice);
-      } else {
-        return new SparseOffHeapVectorValues(fieldEntry, vectorData, bytesSlice);
-      }
-    }
-
-    abstract Bits getAcceptOrds(Bits acceptDocs);
   }
 
   /** Read the nearest-neighbors graph from the index input */
